refactor: add formula module for services
